Switched "refresh installed tools" admin button to use ajax.
When the button is clicked the command to refresh the tools will be sent and the page will refresh after 5 seconds.

Currently a star is displayed to show the page is working since the spinner image is not included in the admin site template. This will be fixed after patching review board to include the spinner image.