Add dependency checking for Tools.
Tools may now provide a 'check_dependencies' method which will be
called when initializing the worker, and refreshing tools. This method
should check the system for any dependencies the tool requires.

If dependencies are missing 'check_dependencies' should return false
and the worker will not listen on the tools queue, or send the tool in
it's list of installed tools.

This changes the Tool API slightly, and some tools may have to be
updated to work with the changes to the Tool class.

The PEP8 tool has been updated to check for the 'pep8' dependency.