Make review settings per tool.
Originally the settings for whether a tool marked a review with
Ship-It, opened issues, or commented on unmodified code were global.
This change makes these settings per tool.

This commit includes changes to both the worker and the extension
which are not backwards compatible (i.e. the new worker will not
work with the old extension, and vice versa).

This commit also requires changes to the database schema which the
included evolutions should provide.