Add cppint and Cppcheck tools
This updates adds support for 2 new workers. cpplint and Cppcheck.
This allows code style to be checked according to the cpplint rules
(various configurable options are available which are not yet exposed
in the Review Bot options). It then checks the code using the static
analysis tool cppcheck. This will raise an Issue if it discovers an
'Error' otherwise just a comment.

Whilst developing this I added a few patches to the core
functionality. So the temp file now has a file extension (where
applicable) - otherwise cpplint and cppcheck refuse to operate on the
file (they want .c/.cpp/.cc etc)

Reviewed at http://reviews.reviewboard.org/r/3579/