Fix Review Bot when SITE_ROOT is used.
When the extension would pass the Review Board instance url to the
workers it was not including SITE_ROOT. This would cause Review Bot to
fail when Review Board was not installed at the root of a domain.

The SITE_ROOT is now included in the generated urls.

Reviewed at http://reviews.reviewboard.org/r/4036/

Closes #6