Update Review Bot to work with Review Board 2.0.
Review Board had some changes that broke Review Bot in a few ways, some
subtle.

First, there are the usual updates for import locations. Django changed
import locations for the urlpattern-related objects, which are now fixed
here.

Then there are more complicated fixes.

Review Bot was unable to log in to Review Board with the provided
session key. That's because Django now checks that the stored auth
backend for a session is in our list of backends. ModelBackend isn't, so
all logins failed. Instead of storing ModelBackend, we now store our
StandardAuthBackend.

After that, everything worked except for publishing. The API resource
handling publishing of reviews was trying to use positional arguments
when fetching an object, but it needed to use keyword arguments. In
previous versions, positional arguments were apparently working, but
this was never intended to be supported.

There's also an update to use our new setup() function instead of the
one that comes with setuptools.

This, along with some changes made to Djblets and Review Board, make
Review Bot functional again.

Reviewed at https://reviews.reviewboard.org/r/5806/