Update Review Bot to take advantage of more modern RB 2.0 features.
Extensions have come a long way in Review Board 2.0. Review Bot no
longer needs to manage its own signal connections or do as much work in
initialize.

I've also updated it to show some better metadata on the extension page,
as far as the name and summary goes. The Author is pointing to the
project instead of just Steven, now that we host it on the 'reviewboard'
GitHub organization. It also will hopefully help take some of the
support load off of just him.

Reviewed at https://reviews.reviewboard.org/r/5809/