Move to a more flexible versioning and package data scheme.
This introduces the same versioning scheme we use in Review Board and
other packages. Version information lives in reviewbotext/__init__.py,
and setup pulls that information in for the package.

To simplify setup.py, the explicitly-defined package data has been
removed, using a MANIFEST.in file instead. This is nicer in that we can
make use of rules like 'recursive-include', instead of having to list
files in each directory explicitly.

Reviewed at https://reviews.reviewboard.org/r/5813/