Add form in admin site for ManualPermissions.
ManualPermissions can now be created, modified, or deleted through the admin
site. I decided to go with a raw ID field for `user` since a drop-down isn't
the best for Review Board instances with many users, but we can change this in
the future if we'd like to.

Testing Done:
- Added, modified, and deleted ManualPermissions successfully.
- Tried entering invalid IDs for the user and local site, and got the expected
  error messages.

Reviewed at https://reviews.reviewboard.org/r/6225/