Don't create a review if there are no comments.
If a Review Bot tool doesn't have anything to say, we don't want to create and
publish an empty review (since doing so would still send an e-mail). This
change fixes it so we only create a review associated with the status update if
there are comments.

Testing done:
Triggered a tool that resulted in zero comments, and checked the resulting
status update. No review was created or associated.

Reviewed at https://reviews.reviewboard.org/r/8533/