Fix up evolutions for Review Bot 1.0.
This change adds a new evolution for people migrating from Review Bot 0.1 or
0.2 to what will be released as 1.0, concerning the removal of the old
`ReviewBotTool` model.

Aditionally, I've decided to remove the evolution that added the timeout field
to the new tool model. Nobody is using this code yet, and so having an extra
evolution for what was really just work-in-progress is silly.

Testing done:
Installed and activated Review Bot 0.1. Upgraded to the master branch and
enabled the extension. Saw that evolutions ran correctly.

Reviewed at https://reviews.reviewboard.org/r/8747/