Add an optional config file for the Review Bot worker.
The worker currently runs without configuration, but some of the new features
we're looking at will require it. In particular, the immediate need is that
PMD, a tool which we'll be adding, does not have a standard installation
mechanism, and often is run via a script from the download directory. The path
to this script has to be configured for the tool to work.

While I was in here, I fixed up the task serialization/accept config to get rid
of a big warning about accepting pickle content (which is insecure and going
away in future versions), and corrected the method of setting the broker URL,
which I hadn't noticed was incorrect because it defaults to localhost.

Testing done:
Ran the worker with and without a config file in place. Saw the expected
output, and that the (future) PMD tool loaded correctly.

Reviewed at https://reviews.reviewboard.org/r/8773/