Update extension and worker setup.py files.
This change updates the setup.py files for the two pieces of Review Bot. The
biggest change is the package names, but this also updates the formatting a bit
and reassigns the author/maintainer fields to be generic Beanbag addresses.

Testing done:
Ran setup.py develop in both packages and saw that things were set up as
expected.

Reviewed at https://reviews.reviewboard.org/r/8748/