Use rich text for Review Bot reviews.
This change enables the use of rich text in comments (which is currently not
used by any tools), and more usefully, uses rich text to improve the format of
the warning when the number of comments posted has been clipped.

Testing done:
Ran a tool that createdmore than 30 comments and saw the resulting body-top was
shown with markdown formatting.

Reviewed at https://reviews.reviewboard.org/r/8763/