Add a PMD tool for Review Bot.
This change adds support for the PMD static analyzer (which is mostly
java-focused). This requires configuration of the path to the executable in
the worker config file, since PMD doesn't have a standard installation method),
and the tool itself can be configured with a list of rulesets to run.

Testing done:
- Posted a big java change and saw that PMD ran and reported the results
  successfully.
- Misconfigured the PMD path and saw that the tool did not list itself as
  available.

Reviewed at https://reviews.reviewboard.org/r/8764/