Update documentation.
This change updates the README for Review Bot, explaining how to install and
configure everything. While I was at it, I updated AUTHORS (making it a simpler
format, since this is a pretty simple package), and added a COPYING file with
the license.

Reviewed at https://reviews.reviewboard.org/r/8749/