Add a flake8 tool.
This adds support for flake8 through Review Bot, which is just a combination of
other, already available tools (pycodestyle and pyflakes).

While I was here I noticed the description for one of the pycodestyle options
was incorrectly referencing the old pep8 tool, and the filename parsing was
fragile in the same way that flake8 would be, so I've fixed that up.

Testing done:
Configured the tool and saw it run successfully.

Reviewed at https://reviews.reviewboard.org/r/8900/