Add support to PMD tool to scan specific files only.
PMD supports multiple languages. But most checks are for java files only as the
other languages are for CPD only or requires manually added checks.

This features allows to increase the performance of review checks as it is not
necessary to run PMD for all files in a diff if only .java has enabled checks.
This avoids but does not fix bug #4633 (which has to do with binary diffs
coming from hg).

This also makes a couple other small changes to PMD:
- Increase timeout as big patches with a lot of files will take some time to be
  finished.
- Fixes conflicting variable name 'f' introduced by changeset
  f95629e2739037bd0e0f289875f710f2da4aa2dd by mistake.

Testing Done:
- Added some binary files with another file extension and saw that it wasn't
  scanned.
- Enabled java extension only and added a java file. Saw that the check found
  correct issues.
- Leave the extension setting empty and saw that it still found correct issues
  in uploaded java patch.
- Used "cpp" as file extension and saw that it ignores java files.

Reviewed at https://reviews.reviewboard.org/r/9530/