Gracefully handle unselected tool.
If a tool was not selected (i.e. user left the dropdown at "---------") when
the user saves the tool, the server would result in an Internal Error. This was
due to the fact that we weren't handling the "empty" case when attempting to
get the data from the widget. Now we handle the empty case, which returns no
data.

Testing Done:
Create new configuration, or edit config:
- No name + no tool -> both error
- Name + no tool -> tool errors

Saw that the proper "This field is required" error was displayed. Selecting a
valid tool makes the error go away.

Bugs closed: 4713

Reviewed at https://reviews.reviewboard.org/r/10064/