Rewrite the README for Review Bot.
This replaces the Review Bot README with a ReST-based version that
summarizes the important parts of Review Bot and defers to the manual
for the rest. It helps keep the important pieces of information in one
place (the manual) and gives viewers looking on GitHub a better idea of
where to start.

Worker and extension-specific READMEs have also been added. These will
be added to the packages in an upcoming change, so that they'll be
included in the source distribution and shown on PyPI.

Testing Done:
Viewed all the rendered READMEs and tested links.

Reviewed at https://reviews.reviewboard.org/r/10109/