Update the metadata for the Review Bot packages.
This adds the new READMEs to the Review Bot extension and worker
packages. It also updates the classifiers to promote the stability level
to Production/Stable, add the Quality Assurance topic, and adds the
Python versions.

Testing Done:
Built the packages and checked the resulting metadata for the README and
the classifiers.

Reviewed at https://reviews.reviewboard.org/r/10110/