EasyFix: Make flake8 an optional dependency under "all"
The new "all" optional dependency target is intended to replicate the old
behavior of installing all python dependencies.

Documentation has been updated to reflect this.

Testing Done:
Since this is just a change for pip dependencies I've done:
- `pip install -e ./bot reviewbot-worker`
- `pip install -e ./bot reviewbot-worker[all]`
To test if it installs and doesn't install flake8 under the default
package and the all target, and according to the logs this is the case.

Reviewed at https://reviews.reviewboard.org/r/10151/