Make pycodestyle and pyflakes an optional dependency under "all"
The documentation incorrectly stated that these tools were "automatically
installed", as they were not dependencies in `setup.py`. These dependencies
were added to the optional 'all' target, and the documentation updated
accordingly.

Testing Done:
Since this is just a change for Python dependencies I've done:
- `pip install -e ./bot`
- `pip install -e ./bot[all]`

To test if it installs and doesn't install pycodestyle and pyflakes under the
default package and the all target, and according to the logs this is the case.

Reviewed at https://reviews.reviewboard.org/r/10218/