Make buildbot an optional dependency under all
The new "all" optional dependency target is intended to replicate the
old behavior of installing all python dependencies.

Documentation has been updated to reflect this.

Testing Done:
Since this is just a change for pip dependencies I've done:
- pip install -e ./bot
- pip install -e ./bot[all]

To test if it installs and doesn't install buildbot under the default
package and the "all" target, and according to the logs this is the
case.

Reviewed at https://reviews.reviewboard.org/r/10156/