Fixed RepositoryTool patching of files in uncreated directories
When a RepositoryTool is applying the review's patches to the cloned
repository, it simply writes using `open(p, 'wb')`. This does not create
directories that the patch implies.

This leads to the write failing with:
`IOError: [Errno 2] No such file or directory: u'dir/file'`

This patch creates directories for each file if the path does not exist
already.

Testing Done:
Replciated the error by posting a review for a file in a new directory,
applied patch, and RepositoryTool successfully runs on patched files.

Reviewed at https://reviews.reviewboard.org/r/10247/