Add fetching of git/hg repositories
If there are a lot of Review Bot workers it can be very expensive to
maintain a repository configuration manually. Since a reviewboard
instance already knows the configuration it is now be possible to
configure multiple reviewboard instances and let the Review Bot fetch
git and hg repositories automatically.

If a repository has a "path" or "mirror_path" that can be access locally
or has a scheme starting with "http" or "git" it will be added to known
repository.

Testing Done:
- Added some repositories in reviewboard with path and mirror_path. Some
  with non-existing local path and mirror_path with "http" scheme.
- Saw that the Review Bot finds all accessible repositories are added
  correctly.
- Tested with and without authentication.

Reviewed at https://reviews.reviewboard.org/r/9502/