Update Review Bot extension configuration to use new admin styling.
This change moves the Review Bot extension configuration over to the new
components for styling in the admin UI. As part of this, I've simplified
the layout a little bit so that the main configuration items are
together in a single fieldset, rather than splitting them into two
labeled ones.

Testing Done:
Verified the appearance and functionality of the extension config.

Reviewed at https://reviews.reviewboard.org/r/10990/