Add back admin site compatibility for Review Board 3.0.
The way that the admin site pages are laid out (in both markup and
styling) is pretty different between Review Board 3.0 and 4.0. The 4.0
way is better in almost every way, but it does make things complicated
for extensions which might want to support both while also having
templates that do more customization than your average form.

This change makes the Review Bot extension detect which version of
Review Board it's on, and render different JS templates with the correct
markup for each.

Testing Done:
Ran the Review Bot extension under Review Board 3.0 and 4.0. Saw that
the UI for configuring the extension looked nice and worked correctly on
both versions.

Reviewed at https://reviews.reviewboard.org/r/11098/