Update the documentation for Review Bot 2.0.
When we released Review Bot 1.0, it was important that we included
information on migrating from Review Bot 0.1/0.2 to 1.0. This is
probably less important today, but certainly the old information
referenced the 1.0 version too heavily and didn't really aid in
upgrading from 1.0 to 2.0.

This change rewrites the installation docs to focus more on generation
installation, cleaning up some parts in the process to reflect how
things work now.

The upgrade information has been split into a new page, which we may add
to down the road if the process changes. There's details on upgrading
both from Review Bot 1.0+, and from the old 0.1/0.2 releases.

The version number of the documentation is now based on the extension
version, instead of hard-coding "1.0".

Some of the new features/tools have been marked with a "added in 2.0"
note, to help make clear what's new.

There's also a fix for the doc8 instructions, which mistakenly
referred to pydocstyle.

Testing Done:
Built the docs. Checked for bad links and spelling/build errors.

Reviewed at https://reviews.reviewboard.org/r/11158/