Work around Chrome's over-eager validation of integration config form.
Chrome loves to test the validation of forms, even going so far as to
run validation on inputs which are not visible to the user. It then
errors out on the console that it can't focus the input. This change
goes through and adds the `disabled` prop to any tool option inputs
which are hidden (because they don't apply to the currently-selected
tool), working around the validation on the client.

Testing Done:
Was able to create an integration config for Review Bot without Chrome
going nuts.

Reviewed at https://reviews.reviewboard.org/r/11184/