FBInfer for Review Bot
This change adds a new tool to Review Bot for running FBInfer. It
supports all of the following build configurations:

* Android
* Apache Ant
* Buck
* C
* CMake
* Gradle with and without a Wrapper
* iOS
* Java
* Make
* Maven
* Objective-C
* XCode

The usability of this tool is largely dependent on the setup of a user's
Review Board environment. Meaning that if a user would like to use
Review Bot to test an Android project. The user must have a working
Android SDK for the Review Bot worker to use. Setups for each particular
environment will be offered in the documentation.

Testing Done:
Each build configuration was tested manually with its own environment
and repository. More information about the tests can be found in Week 3
and Week 4 of my Notion journal: https://bit.ly/351ttRQ.

Reviewed at https://reviews.reviewboard.org/r/11202/