Gofmt Tool for Review Bot
The tool is designed to be a lightweight solution capable of checking
whether a file is correctly formatted or not based on `go fmt` output.
Unfortunately, other go commands such as `go vet` are better suited to
inherit from the RepositoryTool, as they require the ability to check
entire packages to properly test and vet files.

Testing Done:
Manual testing was done to confirm that `go fmt` is working correctly.

Reviewed at https://reviews.reviewboard.org/r/11265/