Cargo Tool for Review Bot
The Cargo Tool utilizes, `cargo clippy` and `cargo test` to lint and
test Rust code for errors that could be potentially expensive. To do
this, the Cargo Tool requires access to the entire repository so that it
can succesfully build the project with the patched code. This is
necessary so that a project can be accurately analyzed.

Testing Done:
Manual testing was done to confirm that `cargo test` is working
correctly. As of now it just creates general comments based on which
tests have failed. It is worth noting, that if the files do not compile
correctly, no comments will be made. Manual testing was done to confirm
that `cargo clippy` creates comments only for patched files.

Reviewed at https://reviews.reviewboard.org/r/11307/