Update package version info to normalize the alpha/beta tags.
This is part of a series of changes throughout our codebases to switch
from "alpha" and "beta" tags to "a" and "b" for the generated package
version. This matches what will actually be generated when we create
packages.

The code and documentation have been cleaned up in the process, and an
extra field for the version number (allowing a `x.y.z.p`) has been added
to match other packages.

Testing Done:
Successfully built packages, and used this code to match the generated
filename.

Reviewed at https://reviews.reviewboard.org/r/11506/