Add some documentation for the Tool class in the Review Bot worker.
The `Tool` class had some important fields that were not explained. For
instance the `version` field, if changed, could break existing
configurations, and should only be done if introducing breaking changes
to the tool.

To avoid problems in the future, and to prepare for some upcoming work,
these fields, and some other docs in the file, have been fleshed out.

Testing Done:
Checked for spelling errors in the docs.

Reviewed at https://reviews.reviewboard.org/r/11510/