Rename the review_board_servers setting to reviewboard_servers.
Review Bot 2.0 introduced a setting for defining the Review Board
servers that Review Bot may perform API requests on. This setting was
named `review_board_servers`, which isn't very consistent with how we
typically identify Review Board in key/filename form. A more consistent
name would be `reviewboard_servers`.

This change updates our code to use the new setting, and to migrate the
old setting to the new one, with a deprecation warning logged.

Documentation has been updated to reference this change, and to document
it in the Upgrading page.

Testing Done:
Manually tested with the old and new setting. Saw a deprecation notice
for the old one, and saw that the setting transitioned over.

Unit tests pass.

Viewed the documentation and checked for bad links, spelling errors, and
build errors.

Reviewed at https://reviews.reviewboard.org/r/11519/