Improve some testing and setup around configuration.
Recent changes made it easier to replace the worker configuration for
the duration of a test, but it wasn't perfect. The new dependency
support mutates the configuration, which can impact other tests.

We're now careful to reset the configuration at the start of each test,
to avoid these sorts of issues. We're also more careful to deep-copy
data into the configuration.

I've also renamed `default_config` to `DEFAULT_CONFIG`, to better convey
it's intended as a constant and not something mutable.

Testing Done:
Unit tests pass.

This resolved some issues hit with upcoming changes.

Reviewed at https://reviews.reviewboard.org/r/11528/